+++ packages/natives/native/index.js
@@ -33,7 +33,7 @@ export const TtyWriter = nativeBindings.TtyWriter;
 
 // functions
 export const __ompInstallTokioRuntime = nativeBindings.__ompInstallTokioRuntime;
-export const __piNativesV18_0_2 = nativeBindings.__piNativesV18_0_2;
+export const __piNativesV18_0_3 = nativeBindings.__piNativesV18_0_3;
 export const astEdit = nativeBindings.astEdit;
 export const astGrep = nativeBindings.astGrep;
 export const astMatch = nativeBindings.astMatch;